+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 🚀 Features
 
 - **playground**:
-  - resizable code editor with border drag controls - by @thuongtruong109 [<samp>(4e2f3)</samp>]()
+  - add quick guide tour - by @thuongtruong109 [<samp>(1f4b1)</samp>]()
+  - resizable code editor with border drag controls - by @thuongtruong109 [<samp>(59e9a)</samp>](https://github.com/thuongtruong109/flashot/commit/59e9af620c24a9516cc9bc428112e3fafe14b72f)
   - add transparent background option - by @thuongtruong109 [<samp>(b66fa)</samp>](https://github.com/thuongtruong109/flashot/commit/b66fa2e26723e38d9261416e837350c8b2c0dbf4)
   - enable word wrap options - by @thuongtruong109 [<samp>(46a18)</samp>](https://github.com/thuongtruong109/flashot/commit/46a18bd436ab01bbd379793b19898ab8a2a65946)
   - adjust width & height - by @thuongtruong109 [<samp>(52d6b)</samp>](https://github.com/thuongtruong109/flashot/commit/52d6be981e1161311ddc88d496b24a0d38b40b07)